EUP-58: Follow up changes for logger changes
- The JM_OUT_DocOutput_wMail class is calling it's base class process
method that was previously calling PluginLog.exitWithStatus(). As
PluginLog.exitWithStatus() is throwing an exception always, the custom
logic for JM_OUT_DocOutput_wMail was in a catch statement. Due to the
removal of PluginLog.exitWithStatus() the code for successful
termination of the base class process method had to be moved outside the
catch block.