EUP-95
-> PMM CN: Only Validated Pre_Receipt Appearing in Safe Desktop >
Receipts Tab

The idea to fix the issue is to adapt the Auto SI plugin to avoid
populating the event level SI for COMM-PHYS deals create with the CN Pre
Receipt template if the Pre Receipt deal is linked to a batch.

The Auto SI process for COMM-PHYS deals based on the CN Pre Receipt
template would be the following:
1.If the deal gets processed to New or Quote, the Auto SI populates the
tran parameter info fields "SI-Phys" and "SI-Phys Internal", potentially
asking the user for input.
2.If the deal gets processed to Validated AND the deal is NOT yet linked
to a batch it is going to behave the same like when processing to new or
quote.
3.If the deal gets processed to Validated AND the deal is linked to a
batch, it is going to pick up the selected SIs form "SI-Phys" and
"SI-Phys Internal" and populate the event info level SI with those
values. No GUI dialog needs to be shown during this step any more.