EUP-124: File extension fix
The Endur core for BO does recognize a limited number of file extensions
excluding XML. As in V17 the Endur core does always change the file
extension of a BO document output type to their default ending
the "Plain Text" now gets the forced extension .txt even in case the
filename pattern requires a .xml file. To circumvent this issue a
new custom logic has been added to rename the generated file on the
filesystem to the defined extension.