EPI-1756: Defect fix in Jackson annotiations
- The Jackson annotations had been setup to require an explicit type
info about which order subtype has to be retrieved. This had been
introduced recently and broke the serialisation of order data for POSTs,
but not for GETs.
The order type is currently being captured an as explicit attribute in
the Order. The explicitly required type annotation has therefore been
removed.